Version 2.1
The start of a major rework of the extruder design to factor into account the wires.

This version changes the main section of the extruder to be wider and to have holes on the top for the wires.

It also has a different Igus Slide Holder that doesn't block the IR Sensor's wires. It does so by converting the slides to be on a vertical instead of horizontal plane.