Designed a compression holder for the Igus Bearings
Compression bearing holders are simply easier to design and print. Also, it uses more sturdy flexures to have a tighter grip on the bearing. Furthermore, this design does not use the ridges the old holder uses. This design purely uses frictional forces to hold the bearing in place. This allows it to be relatively easy to remove when needed.