Redesign of the Extruder
Very rough design for the extruder mount. Instead of trying to squeeze the IR sensor on the mount, I'm focusing on mounting the hotend and the fans. I am also attempting to use an embedded part cooling fan duct instead of a separate part.
This version is still a very rough design that requires wire paths, mounting holes, and a top cover.