Changes to belt tensioner and minor misc changes
Made the non-adjustable belt tensioner to be held by tension instead of by screw.
Removed the extra thickness in the igus holder that initially prevented the bearings from sliding all the way through.

Fixed the size of the IR sensor mounting holes to allow the screw to directly screw into the plastic.